Fix the pawns movements
When moving to the left, Pacman was upside down. It is now moving
properly, but as its eye was not showing correctly, I decided to create
a model in Blender to avoid losing geometry data and having to recreate
it from scratch every time a small change was needed.

There was an issue with the enemy movement as well. It was a bit
sluggish when making turns. Now it will go smoothly anywhere. The speed
up bug when following Pacman in a straight line through teleportation
was also fixed.

YouTube video: https://www.youtube.com/watch?v=waFl619QP1w